+++ src/app/table/table.component.ts
@@ -1,5 +1,7 @@
 import {Component, ViewEncapsulation} from "@angular/core";
 
+import {SettingsService} from "../services/settings.service";
+
 @Component({
   selector: "app-table",
   templateUrl: "./table.component.html",
@@ -7,4 +9,6 @@ import {Component, ViewEncapsulation} from "@angular/core";
   encapsulation: ViewEncapsulation.None
 })
 export class TableComponent{
+  constructor(public settingsService: SettingsService){
+  }
 }