+++ src/app/home-page/home-page.component.ts
@@ -1,9 +1,16 @@
-import {Component} from "@angular/core";
+import {Component, OnInit} from "@angular/core";
+import {Title} from "@angular/platform-browser";
 
 @Component({
   selector: "home-page",
   templateUrl: "./home-page.component.html",
   styleUrls: ["./home-page.component.scss"]
 })
-export class HomePageComponent{
+export class HomePageComponent implements OnInit{
+  constructor(private title: Title){
+  }
+
+  ngOnInit(){
+    this.title.setTitle("BZList");
+  }
 }