+++ src/app/app.module.ts
@@ -2,20 +2,25 @@ import {NgModule} from "@angular/core";
 import {FormsModule} from "@angular/forms";
 import {BrowserModule} from "@angular/platform-browser";
 import {BrowserAnimationsModule} from "@angular/platform-browser/animations";
+import {HttpClientModule} from "@angular/common/http";
 import {ServiceWorkerModule} from "@angular/service-worker";
 
 import {environment} from "../environments/environment";
 
+// modules
 import {AppRoutingModule} from "./app-routing.module";
 import {MaterialModule} from "./material.module";
 
 // firebase
 import {AngularFireModule} from "@angular/fire";
 import {AngularFirestoreModule} from "@angular/fire/firestore";
+import {AngularFireAuthModule} from "@angular/fire/auth";
 import {firebase} from "./firebase";
 
+// services
 import {CookieService} from "ngx-cookie-service";
 import {SettingsService} from "./services/settings.service";
+import {AuthService} from "./services/auth.service";
 
 // pipes
 import {TimeAgoPipe} from "./pipes/time-ago.pipe";
@@ -25,30 +30,34 @@ import {VerboseGameStylePipe} from "./pipes/verbose-game-style.pipe";
 
 // components
 import {AppComponent} from "./app.component";
-import {HomePageComponent} from "./home-page/home-page.component";
-import {HelpPageComponent} from "./help-page/help-page.component";
 import {ServersComponent} from "./servers/servers.component";
 import {NavigationToolbarComponent} from "./navigation-toolbar/navigation-toolbar.component";
-import {ServerPageComponent} from "./server-page/server-page.component";
-import {SettingsPageComponent} from "./settings-page/settings-page.component";
 import {ServerCardComponent} from "./server-card/server-card.component";
 import {ServersTableComponent} from "./servers-table/servers-table.component";
 import {TableComponent} from "./table/table.component";
 import {PlayersTableComponent} from "./players-table/players-table.component";
 
+// pages
+import {ServerPageComponent} from "./server-page/server-page.component";
+import {SettingsPageComponent} from "./settings-page/settings-page.component";
+import {HomePageComponent} from "./home-page/home-page.component";
+import {HelpPageComponent} from "./help-page/help-page.component";
+import {AccountPageComponent} from "./pages/account-page/account-page.component";
+
 @NgModule({
   declarations: [
     AppComponent,
-    HomePageComponent,
-    HelpPageComponent,
     ServersComponent,
     NavigationToolbarComponent,
-    ServerPageComponent,
-    SettingsPageComponent,
     ServerCardComponent,
     ServersTableComponent,
     TableComponent,
     PlayersTableComponent,
+    HomePageComponent,
+    HelpPageComponent,
+    ServerPageComponent,
+    SettingsPageComponent,
+    AccountPageComponent,
     TimeAgoPipe,
     BooleanYesNoPipe,
     NumberZeroNoPipe,
@@ -58,16 +67,15 @@ import {PlayersTableComponent} from "./players-table/players-table.component";
     BrowserModule,
     BrowserAnimationsModule,
     AppRoutingModule,
+    HttpClientModule,
     FormsModule,
     MaterialModule,
     ServiceWorkerModule.register("ngsw-worker.js", {enabled: environment.production}),
     AngularFireModule.initializeApp(firebase),
-    AngularFirestoreModule.enablePersistence()
+    AngularFirestoreModule.enablePersistence(),
+    AngularFireAuthModule
   ],
-  providers: [
-    CookieService,
-    SettingsService
-  ],
+  providers: [CookieService, SettingsService, AuthService],
   bootstrap: [AppComponent]
 })
 export class AppModule {}