+++ src/app/pages/account-page/account-page.component.ts
@@ -1,0 +1,13 @@
+import {Component, OnInit} from "@angular/core";
+
+import {AuthService} from "../../services/auth.service";
+
+@Component({
+  selector: "app-account-page",
+  templateUrl: "./account-page.component.html",
+  styleUrls: ["./account-page.component.scss"]
+})
+export class AccountPageComponent implements OnInit{
+  constructor(public auth: AuthService){
+  }
+}