+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Bleeding Edge
+## 0.3.0 - User accounts 👨‍💼 (2019-4-23)
 
 ### Additions
 - Meta tags